+++ src/crypto/eddsa_key_pair.cpp
@@ -22,6 +22,11 @@ namespace crypto
     return std::make_shared<PublicKeyImpl>(pem);
   }
 
+  EdDSAPublicKeyPtr make_eddsa_public_key(const JsonWebKeyEdDSAPublic& jwk)
+  {
+    return std::make_shared<PublicKeyImpl>(jwk);
+  }
+
   EdDSAKeyPairPtr make_eddsa_key_pair(CurveID curve_id)
   {
     return std::make_shared<KeyPairImpl>(curve_id);
@@ -30,5 +35,10 @@ namespace crypto
   EdDSAKeyPairPtr make_eddsa_key_pair(const Pem& pem)
   {
     return std::make_shared<KeyPairImpl>(pem);
+  }
+
+  EdDSAKeyPairPtr make_eddsa_key_pair(const JsonWebKeyEdDSAPrivate& jwk)
+  {
+    return std::make_shared<KeyPairImpl>(jwk);
   }
 }